+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.1
 
 - first PR by fred
+- collect information about physical property before starting assessment
 
 # 0.0.0 Alpha version